Перенёс все данные карточки в один файл
